delete

+++ src/controller/mem/ReviewDeleteController.java
@@ -7,6 +7,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import dto.XReview;
 import service.face.ReviewService;
@@ -22,7 +23,24 @@ public class ReviewDeleteController extends HttpServlet {
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		System.out.println("/review/delete [GET]");
 		
-		XReview reviewno= reviewService.getReviewNo(req);
+		String param = req.getParameter("review_no");
+		System.out.println(param);
+		int reviewno = 0;
+		if(param != null && !"".equals(param)) {
+			System.out.println(reviewno);
+			reviewno = Integer.parseInt(param);
+			System.out.println(reviewno);
+		} else {
+			System.out.println(reviewno);
+			System.out.println("!!!ERROR!!!");
+			return;
+		}
+		
+		XReview review= reviewService.getReviewDetail(reviewno);
+		HttpSession session = req.getSession();
+		if(!review.getMemId().equals(session.getAttribute("mem_id"))) {
+			return;
+		}
 		
 		reviewService.delete(reviewno);
